+++ CHANGELOG.md
@@ -7,6 +7,8 @@ New features:
 Bug fixes:
 
 * WeakRef now has the same inheritance and methods as MRI's version.
+* Using `Module#const_get` with a scoped argument will now correctly
+  autoload the constant if needed.
 
 # 1.0 RC 5, August 2018